requirements.txt
mcp
starlette
uvicorn
requests
beautifulsoup4
duckduckgo-search
tldextract
Pillow